components/crawlers/rss/requirements.txt
google-cloud-pubsub
feedparser